requirements.txt
flask>=3.0
requests>=2.31
beautifulsoup4>=4.12
openai>=1.0
ddgs>=7.0
dnspython>=2.4
